perf: change way to check users
